requirements.txt
cffi==1.17.1
cryptography==45.0.4
et_xmlfile==2.0.0
greenlet==3.2.3
numpy==2.3.0
openpyxl==3.1.5
pandas==2.3.0
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
